variant.py: extract spec bits into spec.py (#45941)
